+++ analyses/sun-study/src/main.tsx
@@ -1,5 +1,4 @@
 import { render } from 'preact'
 import { App } from './app.tsx'
-import './style.css'
 
 render(<App />, document.getElementById('app')!)
